requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-binance>=1.0.17
pandas>=1.3.0
websocket-client>=1.2.0
twisted>=21.0.0
ta-lib>=0.4.24
numpy>=1.21.0
aiohttp>=3.8.0
pyyaml>=6.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
binance-sdk-derivatives-trading-usds-futures>=2.0.0
apscheduler>=3.10.4
# LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.0.0
# PostgreSQL dependencies
asyncpg>=0.29.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
